have echothread broadcast message via accessing connection pool
